+++ doc/1_overview.qbk
@@ -7,10 +7,19 @@
 
 [section:overview Introduction]
 
-Beast is a cross-platform, header-only C++11 library for low-level HTTP
-and WebSocket protocol programming that uses the consistent networking
-and asynchronous model of __Asio__. The design of the library achieves
-these goals:
+[important
+    Beast is a cross-platform, header-only C++11 library for low-level HTTP
+    and WebSocket protocol programming that uses the consistent network and
+    asynchronous model of __Asio__.
+    
+    Beast is not an HTTP client or HTTP server, but it can be used to
+    build those things. The library is intended to be a foundation upon
+    which new libraries may be built. It is a goal that other architects
+    will create interoperable solutions on top of Beast. The provided
+    example programs shows how a client or server might be built.
+]
+
+The design of the library achieves these goals:
 
 * [*Symmetry.] Interfaces are role-agnostic; the same interfaces can be
 used to build clients, servers, or both.
@@ -92,14 +101,6 @@ Authentication,  mime/multipart encoding, cookies, automatic handling
 of redirects, gzipped transfer encodings, caching, or proxying (to name
 a few) are not directly provided, but nothing stops users from creating
 these features using Beast's HTTP message types.
-
-[important
-  Beast is not an HTTP client or HTTP server, but it can be used to
-  build both. The library is intended to be a foundation upon which
-  new libraries may be built. It is a goal that other architects will
-  build interoperable solutions on top of Beast. The provided example
-  code shows how a client or server may be built.
-]
 
 [heading Credits]
