+++ doc/3_0_core.qbk
@@ -5,17 +5,7 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ]
 
-[section:core Core Concepts]
-
-[block '''
-<informaltable frame="all"><tgroup cols="1"><colspec colname="a"/><tbody><row><entry valign="top"><simplelist>
-  <member><link linkend="beast.core.asio">Working With Asio</link></member>
-  <member><link linkend="beast.core.streams">Stream Concepts</link></member>
-  <member><link linkend="beast.core.buffers">Buffer Concepts</link></member>
-  <member><link linkend="beast.core.async">Asynchronous Utilities</link></member>
-  <member><link linkend="beast.core.tutorial">Writing Composed Operations</link></member>
-</simplelist></entry></row></tbody></tgroup></informaltable>
-''']
+[section:core Library Basics]
 
 A goal of the library is expose implementation primitives in order that
 users may build their own library-like components. These primitives include
@@ -48,6 +38,6 @@ lists these facilities by group, with descriptions.
 [include 3_2_streams.qbk]
 [include 3_3_buffers.qbk]
 [include 3_4_async.qbk]
-[include 3_5_tutorial.qbk]
+[include 3_5_op_tutorial.qbk]
 
 [endsect]
