+++ doc/4_01_primer.qbk
@@ -5,7 +5,7 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ]
 
-[section:primer HTTP Primer]
+[section:primer Protocol Primer]
 
 The HTTP protocol defines the
 [@https://tools.ietf.org/html/rfc7230#section-2.1 client and server roles]:
