+++ doc/4_06_serializer_buffers.qbk
