+++ doc/6_1_streams.qbk
