+++ doc/6_2_connect.qbk
