+++ doc/6_5_messages.qbk
