+++ doc/6_6_control.qbk
