+++ doc/6_7_notes.qbk
