+++ doc/8_0_design.qbk
@@ -7,15 +7,6 @@
 
 [section:design Design Choices]
 
-[block '''
-<informaltable frame="all"><tgroup cols="1"><colspec colname="a"/><tbody><row><entry valign="top"><simplelist>
-  <member><link linkend="beast.design.http_message">HTTP Message Container</link></member>
-  <member><link linkend="beast.design.http_comparison">HTTP Comparison to Other Libraries</link></member>
-  <member><link linkend="beast.design.websocket_zaphoyd">Comparison to Zaphoyd Studios WebSocket++</link></member>
-  <member><link linkend="beast.design.faq">FAQ</link></member>
-</simplelist></entry></row></tbody></tgroup></informaltable>
-''']
-
 The implementations are driven by business needs of cryptocurrency server
 applications (e.g. [@https://ripple.com Ripple]) written in C++. These
 needs were not met by existing solutions so Beast was written from scratch
@@ -59,9 +50,9 @@ start. Other design goals:
 
 * Allow for customizations, if the user needs it.
 
-[include 6_1_http_message.qbk]
-[include 6_2_http_comparison.qbk]
-[include 6_3_websocket_zaphoyd.qbk]
-[include 6_4_faq.qbk]
+[include 8_1_http_message.qbk]
+[include 8_2_http_comparison.qbk]
+[include 8_3_websocket_zaphoyd.qbk]
+[include 8_4_faq.qbk]
 
 [endsect]
