+++ doc/8_4_faq.qbk
@@ -5,7 +5,7 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ]
 
-[section:faq Boost Formal Review FAQ]
+[section:faq FAQ]
 
 To set realistic expectations and prevent a litany of duplicate review
 statements, these notes address the most common questions and comments
