+++ doc/concept/BufferSequence.qbk
@@ -5,9 +5,9 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ]
 
-[section:BufferSequence BufferSequence requirements]
+[section:BufferSequence BufferSequence]
 
-A `BufferSequence` is a type meeting either of the following requirements:
+A [*BufferSequence] is a type meeting either of the following requirements:
 
 * __ConstBufferSequence__
 * __MutableBufferSequence__
