+++ doc/concept/Field.qbk
@@ -5,7 +5,7 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ]
 
-[section:Field Field requirements]
+[section:Field Field]
 
 A [*Field] represents a single HTTP header field/value pair.
