+++ doc/reference.xsl
@@ -1609,7 +1609,7 @@
   <xsl:text>    </xsl:text>
   <xsl:choose>
     <xsl:when test="type = 'class AsyncStream'">
-      <xsl:text>class ``[link beast.ref.streams.AsyncStream [*AsyncStream]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.streams.AsyncStream [*AsyncStream]]``</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class AsyncReadStream'">
       <xsl:text>class __AsyncReadStream__</xsl:text>
@@ -1618,14 +1618,14 @@
       <xsl:text>class __AsyncWriteStream__</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class Body'">
-      <xsl:text>class ``[link beast.ref.Body [*Body]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.Body [*Body]]``</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class BufferSequence'">
-      <xsl:text>class ``[link beast.ref.BufferSequence [*BufferSequence]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.BufferSequence [*BufferSequence]]``</xsl:text>
     </xsl:when>
     <xsl:when test="(type = 'class' or type = 'class...') and declname = 'BufferSequence'">
       <xsl:value-of select="type"/>
-      <xsl:text> ``[link beast.ref.BufferSequence [*BufferSequence]]``</xsl:text>
+      <xsl:text> ``[link beast.concept.BufferSequence [*BufferSequence]]``</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'CompletionHandler' or type = 'class CompletionHandler'">
       <xsl:text>class __CompletionHandler__</xsl:text>
@@ -1634,7 +1634,7 @@
       <xsl:text>class __ConstBufferSequence__</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'DynamicBuffer' or type = 'class DynamicBuffer'">
-      <xsl:text>class ``[link beast.ref.DynamicBuffer [*DynamicBuffer]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.DynamicBuffer [*DynamicBuffer]]``</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'Handler' or type = 'class Handler'">
       <xsl:text>class __Handler__</xsl:text>
@@ -1642,14 +1642,11 @@
     <xsl:when test="declname = 'MutableBufferSequence' or type = 'class MutableBufferSequence'">
       <xsl:text>class __MutableBufferSequence__</xsl:text>
     </xsl:when>
-    <xsl:when test="declname = 'Parser' or type = 'class Parser'">
-      <xsl:text>class ``[link beast.ref.Parser [*Parser]]``</xsl:text>
-    </xsl:when>
     <xsl:when test="declname = 'Stream' or type = 'class Stream'">
-      <xsl:text>class ``[link beast.ref.streams.Stream [*Stream]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.streams.Stream [*Stream]]``</xsl:text>
     </xsl:when>
     <xsl:when test="type = 'class SyncStream'">
-      <xsl:text>class ``[link beast.ref.streams.SyncStream [*SyncStream]]``</xsl:text>
+      <xsl:text>class ``[link beast.concept.streams.SyncStream [*SyncStream]]``</xsl:text>
     </xsl:when>
     <xsl:when test="declname = 'SyncReadStream' or type = 'class SyncReadStream'">
       <xsl:text>class __SyncReadStream__</xsl:text>
