+++ include/beast.hpp
@@ -1,0 +1,19 @@
+//
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
+#ifndef BEAST_HPP
+#define BEAST_HPP
+
+#include <beast/config.hpp>
+
+#include <beast/core.hpp>
+#include <beast/http.hpp>
+#include <beast/version.hpp>
+#include <beast/websocket.hpp>
+#include <beast/zlib.hpp>
+
+#endif
