+++ test/Jamfile
@@ -43,7 +43,7 @@ unit-test http-tests :
     ../extras/beast/unit_test/main.cpp
     http/basic_parser.cpp
     http/buffer_body.cpp
-    http/design.cpp
+    http/doc_http_samples.cpp
     http/dynamic_body.cpp
     http/error.cpp
     http/fields.cpp
